Part Two
- Separate in multiple controller administration part to facilitate the maintenance of the various pages and systems linked to the api of your games.

- Add Guild page on administration
