+++ templates/Britania/admin/includes/_sidebar.html.twig
@@ -93,6 +93,13 @@
     </li>
 
     <li class="nav-item">
+        <a class="nav-link" href="{{path('admin.guilds')}}">
+            <i class="fas fa-users"></i>
+            <span>{{"Liste des guildes"|trans}}</span>
+        </a>
+    </li>
+
+    <li class="nav-item">
         <a class="nav-link" href="{{path('admin.items')}}">
             <i class="fas fa-database"></i>
             <span>{{"Voir la liste des items"|trans}}</span>
