+++ templates/BritaniaR/admin/guild/index.html.twig
@@ -1,0 +1,62 @@
+{% extends settings.get("theme") ~'/admin.html.twig' %}
+
+{% block title %}
+	{{"Liste des guildes"|trans}}
+{% endblock %}
+
+{% block body %}
+	<div class="card shadow mb-4">
+		<div class="card-header py-3">
+			<h6 class="m-0 font-weight-bold text-primary">{{'Liste des guildes'|trans}}</h6>
+		</div>
+		<div class="card-body">
+			<div class="table-responsive">
+
+				<table class="ladder table py-3">
+					<thead>
+						<tr>
+							<th>{{"#"|trans}}</th>
+							<th class="c">{{"Nom"|trans}}</th>
+							<th class="c">{{"Date de fondation"|trans}}</th>
+							<th class="c">{{"Nombre de membres"|trans}}</th>
+							<th class="c">{{"Action"|trans}}</th>
+						</tr>
+					</thead>
+					<tbody>
+						{% for guild in guilds %}
+							<tr>
+								<td class="item_id">
+									{{ guild.guildInfo.Id }}
+								</td>
+								<td class="c">{{ guild.guildInfo.Name }}</td>
+								<td class="c">
+									{{ guild.guildInfo.FoundingDate|date("m/d/Y") }}
+								</td>
+								<td class="c">
+									{{ guild.members|length }}
+								</td>
+								<td>
+									<a href="{{path("admin_guild.get", {"id":guild.guildInfo.Id})}}">{{"Voir la guilde"|trans}}</a>
+								</td>
+							</tr>
+						{% endfor %}
+					</tbody>
+				</table>
+
+				<div
+					class="bg-white p-3">{# {{ knp_pagination_render(items) }} #}
+				</div>
+				{# <nav aria-label="Pagination de la liste des joueurs" class="d-none">
+								                <ul class="pagination p-3 overflow-auto">
+								                    {% for i in 0..total_page %}
+								                    <li class="page-item  {% if page_actuel == i %}active{% endif %}">
+								                        <a class="page-link" data-page="{{i}}" href="{{path('admin.account', {"page":i})}}">{{ i +
+								                            1}}</a>
+								                    </li>
+								                    {% endfor %}
+								                </ul>
+								            </nav> #}
+			</div>
+		</div>
+	</div>
+{% endblock %}
